JDK 21 Virtual Threads
Potential fix for pull request finding
