+++ .github/workflows/maven.yml
@@ -1,10 +1,5 @@
-# This workflow will build a Java project with Maven, and cache/restore any dependencies to improve the workflow execution time
-# For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-java-with-maven
-
-# This workflow uses actions that are not certified by GitHub.
-# They are provided by a third-party and are governed by
-# separate terms of service, privacy policy, and support
-# documentation.
+# Build and test JmDNS on multiple JDKs.
+# Compile requires JDK 21+ (multi-release java21 sources); the shipped jar targets Java 8.
 
 name: Java CI with Maven
 
@@ -16,26 +11,47 @@ on:
 
 jobs:
   build:
-
-    name: JDK ${{ matrix.java }} on ${{ matrix.os }}
+    runs-on: ubuntu-latest
     strategy:
       fail-fast: false
       matrix:
-        os: [windows-latest, ubuntu-latest]
-        java: [17]
-    runs-on: ${{ matrix.os }}
+        java: [8, 11, 17, 21]
+
     steps:
     - uses: actions/checkout@v4
-    - name: Set up JDK ${{ matrix.java }}
+
+    - name: Set up JDKs (21 for compile + ${{ matrix.java }} for tests)
       uses: actions/setup-java@v4
       with:
-        java-version: ${{ matrix.java }}
-        distribution: 'temurin'
+        distribution: temurin
+        java-version: |
+          21
+          ${{ matrix.java }}
         cache: maven
-    - name: Build with Maven
-      run: mvn -B package --file pom.xml
 
-    # Optional: Uploads the full dependency graph to GitHub to improve the quality of Dependabot alerts this repository can receive
+    - name: Build and test on JDK ${{ matrix.java }}
+      run: |
+        if [ "${{ matrix.java }}" = "8" ]; then
+          mvn -B package -DskipTests --file pom.xml
+        else
+          mvn -B verify -Dtest.jdk.version=${{ matrix.java }} --file pom.xml
+        fi
+
+    - name: Verify Java 8 bytecode and multi-release layout
+      if: matrix.java == '8'
+      run: |
+        JAR=$(ls target/jmdns-*.jar | head -1)
+        echo "Verifying $JAR"
+        jar tf "$JAR" | grep -q 'META-INF/versions/21/javax/jmdns/impl/util/ExecutorServiceFactory.class'
+        javap -classpath "$JAR" javax.jmdns.impl.util.ExecutorServiceFactory | grep -q 'major version: 52'
+        ! javap -classpath "$JAR" javax.jmdns.impl.util.ExecutorServiceFactory | grep -q 'newVirtualThreadPerTaskExecutor'
+
+    - name: Verify virtual-thread implementation on JDK 21
+      if: matrix.java == '21'
+      run: |
+        JAR=$(ls target/jmdns-*.jar | head -1)
+        javap -classpath "$JAR" javax.jmdns.impl.util.ExecutorServiceFactory | grep -q 'newVirtualThreadPerTaskExecutor'
+
     - name: Update dependency graph
-      if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+      if: github.event_name == 'push' && github.ref == 'refs/heads/main' && matrix.java == '21'
       uses: advanced-security/maven-dependency-submission-action@571e99aab1055c2e71a1e2309b9691de18d6b7d6
