+++ .settings/org.eclipse.core.resources.prefs
@@ -1,6 +1,6 @@
-#Sat Nov 19 09:18:33 PST 2011
 eclipse.preferences.version=1
 encoding//src/main/java=UTF-8
+encoding//src/main/java21=UTF-8
 encoding//src/main/resources=UTF-8
 encoding//src/test/java=UTF-8
 encoding//src/test/resources=UTF-8
