+++ pom.xml
@@ -3,7 +3,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>org.jmdns</groupId>
 	<artifactId>jmdns</artifactId>
-	<version>3.6.4-SNAPSHOT</version>
+	<version>4.0.0-SNAPSHOT</version>
 	<name>JmDNS</name>
 	<packaging>jar</packaging>
 
@@ -54,9 +54,11 @@ The project was originally started in December 2002 by Arthur van Hoff at Strang
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<maven.build.timestamp.format>yyyy-MM-dd HH:mm</maven.build.timestamp.format>
-		<jvm.version>8</jvm.version>
-		<javase.version>1.8</javase.version>
-		<mockito.version>4.11.0</mockito.version>
+		<!-- Main sources: Java 8 bytecode; JDK 21+ sources in src/main/java21 (multi-release). -->
+		<maven.compiler.release>8</maven.compiler.release>
+		<!-- Mockito 5.x and modern JUnit require Java 11+ at test compile/runtime. -->
+		<maven.compiler.testRelease>11</maven.compiler.testRelease>
+		<mockito.version>5.12.0</mockito.version>
 	</properties>
 
 	<dependencies>
@@ -85,7 +87,7 @@ The project was originally started in December 2002 by Arthur van Hoff at Strang
 		</dependency>
 		<dependency>
 			<groupId>org.mockito</groupId>
-			<artifactId>mockito-inline</artifactId>
+			<artifactId>mockito-core</artifactId>
 			<version>${mockito.version}</version>
 			<scope>test</scope>
 		</dependency>
@@ -101,12 +103,24 @@ The project was originally started in December 2002 by Arthur van Hoff at Strang
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
-				<version>3.13.0</version>
-				<configuration>
-					<source>${jvm.version}</source>
-					<target>${jvm.version}</target>
-					<release>${jvm.version}</release>
-				</configuration>
+				<version>3.15.0</version>
+				<executions>
+					<!-- default-compile uses maven.compiler.release (8) for src/main/java -->
+					<execution>
+						<id>compile-java-21</id>
+						<phase>compile</phase>
+						<goals>
+							<goal>compile</goal>
+						</goals>
+						<configuration>
+							<release>21</release>
+							<compileSourceRoots>
+								<compileSourceRoot>${project.basedir}/src/main/java21</compileSourceRoot>
+							</compileSourceRoots>
+							<multiReleaseOutput>true</multiReleaseOutput>
+						</configuration>
+					</execution>
+				</executions>
 			</plugin>
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
@@ -123,24 +137,21 @@ The project was originally started in December 2002 by Arthur van Hoff at Strang
 				</executions>
 				<configuration>
 					<attach>true</attach>
+					<classifier>sources</classifier>
 				</configuration>
 			</plugin>
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-surefire-plugin</artifactId>
 				<version>3.5.0</version>
+				<configuration>
+					<argLine>-XX:+EnableDynamicAgentLoading</argLine>
+				</configuration>
 			</plugin>
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-jar-plugin</artifactId>
-				<version>2.6</version>
-				<executions>
-					<execution>
-						<goals>
-							<goal>jar</goal>
-						</goals>
-					</execution>
-				</executions>
+				<version>3.4.2</version>
 				<configuration>
 					<archive>
 						<addMavenDescriptor />
@@ -156,6 +167,7 @@ The project was originally started in December 2002 by Arthur van Hoff at Strang
 						<manifestEntries>
 							<Implementation-Date>${maven.build.timestamp}</Implementation-Date>
 							<Automatic-Module-Name>javax.jmdns</Automatic-Module-Name>
+							<Multi-Release>true</Multi-Release>
 						</manifestEntries>
 					</archive>
 				</configuration>
@@ -163,7 +175,7 @@ The project was originally started in December 2002 by Arthur van Hoff at Strang
 			<plugin>
 				<groupId>org.apache.felix</groupId>
 				<artifactId>maven-bundle-plugin</artifactId>
-				<version>2.5.4</version>
+				<version>5.1.9</version>
 				<extensions>true</extensions>
 				<configuration>
 					<manifestLocation>${project.build.outputDirectory}/META-INF</manifestLocation>
@@ -173,7 +185,7 @@ The project was originally started in December 2002 by Arthur van Hoff at Strang
 						<Export-Package>javax.jmdns</Export-Package>
 						<Private-Package>javax.jmdns.impl,javax.jmdns.impl.*,com.strangeberry.*,samples</Private-Package>
 						<Import-Package>!javax.swing.*,org.slf4j</Import-Package>
-						<Bundle-RequiredExecutionEnvironment>JavaSE-${javase.version}</Bundle-RequiredExecutionEnvironment>
+						<Bundle-RequiredExecutionEnvironment>JavaSE-1.8</Bundle-RequiredExecutionEnvironment>
 						<Bundle-Vendor>jmdns.org</Bundle-Vendor>
 						<Bundle-ClassPath>.</Bundle-ClassPath>
 					</instructions>
@@ -189,27 +201,26 @@ The project was originally started in December 2002 by Arthur van Hoff at Strang
 				</executions>
 			</plugin>
 			<plugin>
-			    <groupId>org.apache.maven.plugins</groupId>
-			    <artifactId>maven-javadoc-plugin</artifactId>
-			    <version>3.8.0</version>
-			    <configuration>
-			        <outputDirectory>${project.build.directory}/javadoc</outputDirectory>
-			        <source>${jvm.version}</source>
-			        <failOnError>false</failOnError>
-			        <failOnWarnings>false</failOnWarnings>
-			        <doclint>none</doclint>
-			    </configuration>
-			    <executions>
-			        <execution>
-			            <id>attach-javadocs</id>
-           			 	<goals>
-		            	    <goal>jar</goal>
-        			    </goals>
-			            <configuration>
-            			    <additionalparam>${javadoc.opts}</additionalparam>
-			            </configuration>
-			        </execution>
-			    </executions>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-javadoc-plugin</artifactId>
+				<version>3.10.0</version>
+				<configuration>
+					<outputDirectory>${project.build.directory}/javadoc</outputDirectory>
+					<failOnError>false</failOnError>
+					<failOnWarnings>false</failOnWarnings>
+					<doclint>none</doclint>
+				</configuration>
+				<executions>
+					<execution>
+						<id>attach-javadocs</id>
+						<goals>
+							<goal>jar</goal>
+						</goals>
+						<configuration>
+							<additionalparam>${javadoc.opts}</additionalparam>
+						</configuration>
+					</execution>
+				</executions>
 			</plugin>
 			<plugin>
 				<groupId>com.mycila</groupId>
@@ -265,6 +276,27 @@ The project was originally started in December 2002 by Arthur van Hoff at Strang
 
 	<profiles>
 		<profile>
+			<id>ci-test-jdk</id>
+			<activation>
+				<property>
+					<name>test.jdk.version</name>
+				</property>
+			</activation>
+			<build>
+				<plugins>
+					<plugin>
+						<groupId>org.apache.maven.plugins</groupId>
+						<artifactId>maven-surefire-plugin</artifactId>
+						<configuration>
+							<jdkToolchain>
+								<version>${test.jdk.version}</version>
+							</jdkToolchain>
+						</configuration>
+					</plugin>
+				</plugins>
+			</build>
+		</profile>
+		<profile>
 			<id>release-sign-artifacts</id>
 			<activation>
 				<property>
@@ -290,17 +322,6 @@ The project was originally started in December 2002 by Arthur van Hoff at Strang
 					</plugin>
 				</plugins>
 			</build>
-		</profile>
-		<!-- We need this profile in order to set '-Xdoclint:none' as a project property which will be used later by maven-javadoc-plugin as an 'additionalparam' to be passed to the javadoc.exe. -->
-		<!-- This option will be used only if the JDK version is 1.8 or higher. Earlier versions of javadoc.exe does not accept this option. -->
-		<profile>
-			<id>doclint-java8-disable</id>
-			<activation>
-				<jdk>[1.8,)</jdk>
-			</activation>
-			<properties>
-				<javadoc.opts>-Xdoclint:none</javadoc.opts>
-			</properties>
 		</profile>
 	</profiles>
 </project>
